# docs/source/specs/v2/openapi.yaml
openapi: 3.0.0
info:
  title: Role Based Access Control for Kessel Project
  version: v2
...
paths:
...
  /workspaces/{id}/:
    get:
      operationId: Workspaces_read
      summary: Get a workspace in tenant
      description: Get a workspace in tenant
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identification
          schema:
            $ref: '#/components/schemas/UUID'
        - name: include_ancestry
          in: query
          required: false
          description: When true, the response will include the ancestry of the workspace.
          schema:
            type: boolean
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Workspaces.ReadWorkspaceResponse'
                  - $ref: '#/components/schemas/Workspaces.ReadWorkspaceWithAncestryResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 401
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
        '403':
          description: Access is forbidden.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problems.Problem403'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problems.Problem404'
        '500':
          description: Server error
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 500
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
      tags:
        - Workspaces
    put:
      operationId: Workspaces_update
      summary: Edit the workspace name or description
      description: Edit a workspace
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identification
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspaces.UpdateWorkspaceResponse'
              example:
                _: 200
                id: e4277742-b91c-43f1-a185-b827e8574345
                parent_id: c1f729e2-3e2b-4f9e-b247-a4b568393e11
                type: standard
                name: Alpha Workspace
                description: Changing description of alpha workspace to something new.
                created: '2024-08-04T12:00:00Z'
                modified: '2024-08-04T12:00:00Z'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 400
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
        '401':
          description: Access is unauthorized.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 401
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
        '403':
          description: Access is forbidden.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problems.Problem403'
        '500':
          description: Server error
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 500
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
      tags:
        - Workspaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Workspaces.UpdateWorkspaceRequest'
            example:
              parent_id: c1f729e2-3e2b-4f9e-b247-a4b568393e11
              name: Alpha Workspace
              description: Changing description of alpha workspace to something new.
    patch:
      operationId: Workspaces_patch
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identification
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspaces.PatchWorkspaceResponse'
              example:
                _: 200
                id: e4277742-b91c-43f1-a185-b827e8574345
                parent_id: c1f729e2-3e2b-4f9e-b247-a4b568393e11
                type: standard
                name: Alpha Workspace - closed
                description: Create a standard workspace.
                created: '2024-08-04T12:00:00Z'
                modified: '2024-08-04T12:00:00Z'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 400
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
        '401':
          description: Access is unauthorized.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 401
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
        '403':
          description: Access is forbidden.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problems.Problem403'
        '500':
          description: Server error
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 500
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
      tags:
        - Workspaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Workspaces.PatchWorkspaceRequest'
            example:
              name: Alpha Workspace - closed
              description: Create a standard workspace.
    delete:
      operationId: Workspaces_delete
      summary: Delete the workspace
      description: Delete a workspace
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identification
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problems.Workspace.Problem400WorkspaceNotEmpty'
        '401':
          description: Access is unauthorized.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 401
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
        '403':
          description: Access is forbidden.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problems.Problem403'
        '500':
          description: Server error
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 500
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
      tags:
        - Workspaces
  /workspaces/{id}/move/:
    post:
      operationId: Workspaces_move
      summary: Move a workspace to a new parent.
      description: Move a workspace to a new parent.
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identification of the workspace to move
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspaces.MoveWorkspaceResponse'
              example:
                id: e4277742-b91c-43f1-a185-b827e8574345
                parent_id: f0e1d2c3-b4a5-4678-9a1b-c2d3e4f5a6b7
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problems.Problem400'
        '401':
          description: Access is unauthorized.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 401
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
        '403':
          description: Access is forbidden.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problems.Problem403'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problems.Problem404'
        '500':
          description: Server error
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Problems.ProblemType'
                  status:
                    type: number
                    enum:
                      - 500
                  title:
                    type: string
                  detail:
                    type: string
                  instance:
                    type: string
                    format: uri
      tags:
        - Workspaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Workspaces.MoveWorkspaceRequest'
            example:
              parent_id: c1f729e2-3e2b-4f9e-b247-a4b568393e11
components:
  schemas:
...
    Problems.Problem403:
      type: object
      required:
        - type
        - title
      properties:
        type:
          type: string
          enum:
            - http://project-kessel.org/problems/insufficient-permission
        title:
          type: string
          enum:
            - You do not have permission to perform this action.
      allOf:
        - type: object
          properties:
            type:
              $ref: '#/components/schemas/Problems.ProblemType'
            status:
              type: number
              enum:
                - 403
            title:
              type: string
            detail:
              type: string
            instance:
              type: string
              format: uri
    Problems.Problem404:
      type: object
      required:
        - title
        - detail
      properties:
        title:
          type: string
          enum:
            - Resource was not found
        detail:
          type: string
          example: Not found
      allOf:
        - type: object
          properties:
            type:
              $ref: '#/components/schemas/Problems.ProblemType'
            status:
              type: number
              enum:
                - 404
            title:
              type: string
            detail:
              type: string
            instance:
              type: string
              format: uri
    Problems.ProblemType:
      type: string
      enum:
        - http://project-kessel.org/problems/insufficient-permission
    Problems.Workspace.Problem400WorkspaceNotEmpty:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          enum:
            - Unable to delete due to workspace dependencies
      allOf:
        - type: object
          properties:
            type:
              $ref: '#/components/schemas/Problems.ProblemType'
            status:
              type: number
              enum:
                - 400
            title:
              type: string
            detail:
              type: string
            instance:
              type: string
              format: uri
...
    UUID:
      type: string
      format: uuid
...
    Workspaces.MoveWorkspaceResponse:
      type: object
      required:
        - id
        - parent_id
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/UUID'
          default: Workspace ID of moved workspace
        parent_id:
          allOf:
            - $ref: '#/components/schemas/UUID'
          default: New parent ID of moved workspace
      example:
        _: 200
        id: e4277742-b91c-43f1-a185-b827e8574345
        parent_id: f0e1d2c3-b4a5-4678-9a1b-c2d3e4f5a6b7
...
    Workspaces.PatchWorkspaceRequest:
      type: object
      properties:
        name:
          type: string
          description: Workspace A
          default: Workspace A
        description:
          type: string
          description: Description of Workspace A
          default: Description of Workspace A
      example:
        name: Alpha Workspace - closed
        description: Create a standard workspace.
    Workspaces.PatchWorkspaceResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/Workspaces.Workspace'
      example:
        _: 200
        id: e4277742-b91c-43f1-a185-b827e8574345
        parent_id: c1f729e2-3e2b-4f9e-b247-a4b568393e11
        type: standard
        name: Alpha Workspace - closed
        description: Create a standard workspace.
        created: '2024-08-04T12:00:00Z'
        modified: '2024-08-04T12:00:00Z'
    Workspaces.ReadWorkspaceResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/Workspaces.Workspace'
      example:
        _: 200
        id: e4277742-b91c-43f1-a185-b827e8574345
        parent_id: c1f729e2-3e2b-4f9e-b247-a4b568393e11
        type: standard
        name: Alpha Workspace
        description: Create a standard workspace.
        created: '2024-08-04T12:00:00Z'
        modified: '2024-08-04T12:00:00Z'
    Workspaces.ReadWorkspaceWithAncestryResponse:
      type: object
      required:
        - ancestry
      properties:
        ancestry:
          type: array
          items:
            $ref: '#/components/schemas/Workspaces.WorkspaceAncestor'
      allOf:
        - $ref: '#/components/schemas/Workspaces.ReadWorkspaceResponse'
      example:
        ancestry:
          - id: c1f729e2-3e2b-4f9e-b247-a4b568393e11
            parent_id: a1b2c3d4-e5f6-47a8-b9c0-d1e2f3a4b5c6
            name: Default Workspace
        _: 200
        id: e4277742-b91c-43f1-a185-b827e8574345
        parent_id: c1f729e2-3e2b-4f9e-b247-a4b568393e11
        type: standard
        name: Alpha Workspace
        description: Create a standard workspace.
        created: '2024-08-04T12:00:00Z'
        modified: '2024-08-04T12:00:00Z'
    Workspaces.UpdateWorkspaceRequest:
      type: object
      required:
        - name
        - description
        - parent_id
      properties:
        name:
          type: string
          description: Workspace A
          default: Workspace A
        description:
          type: string
          description: Description of Workspace A
          default: Description of Workspace A
        parent_id:
          allOf:
            - $ref: '#/components/schemas/UUID'
          description: Must match current parent_id value. Not updatable directly on the workspace.
          default: Parent ID of Workspace A
      example:
        parent_id: c1f729e2-3e2b-4f9e-b247-a4b568393e11
        name: Alpha Workspace
        description: Changing description of alpha workspace to something new.
    Workspaces.UpdateWorkspaceResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/Workspaces.Workspace'
      example:
        _: 200
        id: e4277742-b91c-43f1-a185-b827e8574345
        parent_id: c1f729e2-3e2b-4f9e-b247-a4b568393e11
        type: standard
        name: Alpha Workspace
        description: Changing description of alpha workspace to something new.
        created: '2024-08-04T12:00:00Z'
        modified: '2024-08-04T12:00:00Z'
...
servers:
  - url: https://console.redhat.com/{basePath}
    description: Production Server
    variables:
      basePath:
        default: api/rbac/v2
  - url: https://console.stage.redhat.com/{basePath}
    description: Stage Server
    variables:
      basePath:
        default: api/rbac/v2
  - url: https://localhost:{port}/{basePath}
    description: Development Server
    variables:
      basePath:
        default: api/rbac/v2
      port:
        default: '8000'